Fixed compilation for 2.5
    0dfbed1df6f9d5b4725546161a9666a04d7c1d44
Fixes to compile with 2.5 kernel
    50de3b0f87db6af588d835f0922c2fc6542958dc
Commented broken assert
    a1212e75759f43bd64f38004c91c19ebc6111585
- define constants. - do the hard reset of codec.  let me know any problems on via823x by this change. - clean up the midi init code. - show the revision number in hex.
    f6e93a27cfe2e916d7a3104946f8737d7332530a
- don't touch the modem id during the first reset wait. it may lead to hang-up on es1968+ak4540.
    ee50ef35bdeb441078000c531b6fb98b350c1237
Fixed comment typo
    03fea06b1c8a1c679ffb225756f672e9f05412b0
Fixed error
    19e02bcdf21086777abe0970267c77fd869cc933
Fixed code syntax
    9ed5018e0f158a5d77338f46a63efaec9d4badd6
Fixed typo
    72b8329f5766a583f22c8765ad849e513413b2cc
Fixed compilation error
    53f93adb486fe24747a36c4503e8c6b535c8623d
Implemented fasync callback
    ae95f53d5644778c135c41fdbc82e20f50ab0c62
- don't pass null pointer as the compatible handler (for hammer). call ioctl callback directly.
    4e1edd21f73391007e38493e6af51a6e68c79321
- fixed compile warnings.
    56a2c8698970de8bee13bb23e2cbceb34fe7ce52
- fixed compile warning.
    0f6f00e5b9f07ab86183a9982f06107b8176cf77
- fixed to reset set_fs().
    070844c464870030d81d85814c4dcda4048c8b22

+++ sound/pci/cs46xx/cs46xx_lib.c
@@ -2797,7 +2797,7 @@ static long snd_cs46xx_io_read(snd_info_entry_t *entry, void *file_private_data,
 	snd_cs46xx_region_t *region = (snd_cs46xx_region_t *)entry->private_data;
 	
 	size = count;
-	if (file->f_pos + size > region->size)
+	if (file->f_pos + (size_t)size > region->size)
 		size = region->size - file->f_pos;
 	if (size > 0) {
 		char *tmp;
